backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf_extra_fields